Some cleanup of the new CLLI code
Small changes to the new CLLI code recently added in
https://github.com/AOMediaCodec/libavif/pull/1194.

Document the CLLI metadata read and write support in the CHANGELOG.

Print uint16_t (unsigned short) values using the %hu format.

Put the clli member next to the other color-related members in the
avifImage struct.

+++ apps/shared/avifutil.c
@@ -122,7 +122,7 @@ static void avifImageDumpInternal(const avifImage * avif, uint32_t gridCols, uin
     }
     printf(" * Progressive    : %s\n", avifProgressiveStateToString(progressiveState));
     if (avif->clli.maxCLL > 0 || avif->clli.maxPALL > 0) {
-        printf(" * CLLI           : %u, %u\n", (uint32_t)avif->clli.maxCLL, (uint32_t)avif->clli.maxPALL);
+        printf(" * CLLI           : %hu, %hu\n", avif->clli.maxCLL, avif->clli.maxPALL);
     }
 }
